compilers: Move ArmCompiler to a mixin module
