+++ mesonbuild/compilers/cpp.py
@@ -30,7 +30,6 @@ from .compilers import (
     IntelGnuLikeCompiler,
     IntelVisualStudioLikeCompiler,
     PGICompiler,
-    ArmCompiler,
     ArmclangCompiler,
     Compiler,
     VisualStudioLikeCompiler,
@@ -38,6 +37,7 @@ from .compilers import (
 from .c_function_attributes import CXX_FUNC_ATTRIBUTES, C_FUNC_ATTRIBUTES
 from .mixins.clike import CLikeCompiler
 from .mixins.ccrx import CcrxCompiler
+from .mixins.arm import ArmCompiler
 
 
 def non_msvc_eh_options(eh, args):
